fix tests
Apparently the 'Chilly Bin' step was on top of the new story form and this did not make the tests pass.
Another thing that was happening was that maybe the spinner was not hidden when capybara tries to click on 'Add story' button and the new story form doesn't appear.